% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{convert_to}
\alias{convert_to}
\alias{convert_to.comcat_url}
\title{Change the method in a comcat_url object}
\usage{
convert_to(u, ...)

\method{convert_to}{comcat_url}(u, to, starttime, endtime,
  verbose = TRUE, to_pos = FALSE, ...)
}
\arguments{
\item{u}{\code{\link{comcat_url}} object}

\item{...}{additional arguments}

\item{to}{query-method to convert to; currently only \code{'count'} and \code{'query'}
are supported, and if this is left blank the function
will choose \code{'count'} if the current query-method is neither of the supported options,
or choose the opposite value if the current method is one or the other of the
supported options.}

\item{starttime}{start time of the query; this will supersede and
value already found in \code{u}}

\item{endtime}{end time of the query; this will supersede and
value already found in \code{u}}

\item{verbose}{logical; should messages be displayed?}

\item{to_pos}{logical; if \code{starttime} and/or \code{endtime}
are supplied, should they be converted to the necessary format?}
}
\value{
\code{\link{comcat_url}} object
}
\description{
Change the method in a comcat_url object
}
\examples{
u0 <- make_comcat_url(method='query')
uc <- convert_to(u0) # convert query to count
u <- convert_to(uc) # restore query
all.equal(u0,u)
}
