% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_helpers.R
\name{event_info}
\alias{event_info}
\alias{event_info.default}
\title{Return event information}
\usage{
event_info(id, ...)

\method{event_info}{default}(id, ...)
}
\arguments{
\item{id}{character; event identifier (i.e., \code{'eventid'} in \code{\link{make_comcat_url}}); this
can be a vector of identifiers}

\item{...}{additional arguments to \code{\link{make_comcat_url}}}
}
\value{
a named list of information for each \code{'id'} should it exist; the information for each event is given as
a \code{\link{tibble}}; result will be empty for any event id that is bad or returned no information
}
\description{
Return event information
}
\details{
Note that this issues a \code{\link{comcat_query}} for each ID, and so could
take considerable time as the number of identifiers grows large.
}
\examples{
\dontrun{
event_info(id="us20005awl")
event_info(id=c("us20005awl","us1000gf8u"))
event_info(id="something bogus")
}
}
