% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_comcat.R
\name{make_adaptive_comcat_url}
\alias{make_adaptive_comcat_url}
\alias{comcat_query.comcat_url_list}
\title{Adaptively adjust ComCat query based on result limits}
\usage{
make_adaptive_comcat_url(..., n_segs = NULL, refine = TRUE, verbose = TRUE)

\method{comcat_query}{comcat_url_list}(x, ...)
}
\arguments{
\item{...}{parameters passed to \code{\link{make_comcat_url}}}

\item{n_segs}{integer; the number of segments to split if
the result is over the 20k limit (enforced by ComCat webservices)}

\item{refine}{logical; if the time period is broken up, check for
the need to create subwindows. Presently refinement is only once,
so results are not (yet) guaranteed.}

\item{verbose}{logical; should messages be printed?}

\item{x}{object with class \code{'comcat_url_list'}, e.g. from \code{\link{make_adaptive_comcat_url}}}
}
\value{
An object of class \code{'comcat_url_list'}
}
\description{
Adaptively adjust ComCat query based on result limits
}
\examples{
\dontrun{
aq <- make_adaptive_comcat_url(starttime='2016-01-01', endtime='2016-04-01')
}
}
\seealso{
\code{\link{comcat_query}}
}
