% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_comcat.R
\name{make_comcat_url}
\alias{make_comcat_url}
\alias{comcat_query}
\alias{comcat_query.comcat_url}
\alias{comcat_query.default}
\title{Search ComCat for basic earthquake data}
\usage{
make_comcat_url(starttime = NULL, endtime = NULL,
  updatedafter = NULL, search_box = NULL, search_circle = NULL,
  mindepth = NULL, maxdepth = NULL, minmagnitude = NULL,
  maxmagnitude = NULL, catalog = NULL, contributor = NULL,
  reviewstatus = NULL, minmmi = NULL, maxmmi = NULL, mincdi = NULL,
  maxcdi = NULL, minfelt = NULL, eventtype = "earthquake",
  alertlevel, producttype, method = NULL, limit = NULL,
  orderby = NULL, format = NULL, ...)

comcat_query(x, ...)

\method{comcat_query}{comcat_url}(x, verbose = TRUE, ...)

\method{comcat_query}{default}(...)
}
\arguments{
\item{starttime, endtime, updatedafter}{time window parameters}

\item{search_box}{list containing \code{minlatitude,maxlatitude,minlongitude,maxlongitude} for lat-lon box parameters;
note that if this and \code{search_circle} are specificied, the code will throw and error.}

\item{search_circle}{list containing \code{latitude,longitude,maxradius,maxradiuskm} for circle search parameters;
note that if this and \code{search_box} are specificied, the code will throw and error.}

\item{mindepth, maxdepth}{depth search parameters}

\item{minmagnitude, maxmagnitude}{magnitude search parameters}

\item{catalog, contributor}{catalog and contributor specifications; omitting them gets the "preferred" solution}

\item{reviewstatus}{review status defaults to all or can be automatic or reviewed only}

\item{minmmi, maxmmi}{Minimum and Maximum values for Maximum Modified Mercalli Intensity reported by ShakeMap.}

\item{mincdi, maxcdi}{Minimum and Maximum values for Maximum Community Determined Intensity reported by Did You Feel It? forms (DYFI).}

\item{minfelt}{limit to events with this many DYFI responses}

\item{eventtype}{the type of event; if \code{eventtype='earthquake'} non-earthquakes
are removed from the search. \emph{Warning: there are numerous options (see
\url{https://earthquake.usgs.gov/fdsnws/event/1/application.json}, and no arg-checking is done here.}}

\item{alertlevel}{limit to events with a specific pager alert; checking done if not \code{\link{missing}}}

\item{producttype}{limits to events with specific product types available; checking done if not \code{\link{missing}}}

\item{method}{query or count etc...}

\item{limit}{integer; the maximum number of events to return -- presently it cannot exceed 20k}

\item{orderby}{orering: time-asc, etc...}

\item{format}{csv, etc...}

\item{...}{additional arguments}

\item{x}{object}

\item{verbose}{logical; should the query by shown?}
}
\value{
A url with all non-NULL parameters that are specificied; this modifies
the \code{\link[httr]{url}} object with an additional attribute for format, making it
an object with \code{\link{'comcat_url'}} inheritance.
}
\description{
Searches the ComCat database for basic hypocentral data for earthquakes
 including origin time, location, magnitude, etc.
}
\examples{
make_comcat_url()
make_comcat_url(method='count')
make_comcat_url(method='count', starttime='2011-03-09', endtime='2011-03-11') # prior to Tohoku M9
make_comcat_url(method='count', starttime='2011-03-11', endtime='2011-03-13') # including Tohoku M9
make_comcat_url(method='count', starttime=Sys.time()-30*60)
u <- make_comcat_url(search_circle = list(latitude=32,longitude=-120,maxradiuskm=400))

all.equal(u, httr::build_url(httr::parse_url(u)), check.attributes = FALSE)

\dontrun{
eqs_raw <- readr::read_csv(u)
eqs <- comcat_query(u)
all.equal(eqs, eqs_raw)

# Plot the earthquakes on a map, scaling them by a proxy for seismic moment
plot(latitude ~ longitude, eqs, cex=0.2+scale(10**mag, center=FALSE))
summary(eqs)
}

# Find out values for different parameters:
app <- make_comcat_url(method="application.json")
\dontrun{
app_values <- comcat_query(app)
str(app_values)
}

}
\references{
\url{https://earthquake.usgs.gov/fdsnws/event/1/},
 \url{https://earthquake.usgs.gov/earthquakes/feed/v1.0/csv.php}
}
\seealso{
\link[httr]{build_url} and \link{comcat_hypo}

\code{\link{adaptive_comcat_query}}
}
