% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_comcat.r
\name{midtime}
\alias{midtime}
\title{Calculate the midpoint time between two times}
\usage{
midtime(early, late)
}
\arguments{
\item{early}{datetime}

\item{late}{datetime}
}
\value{
datetime
}
\description{
Calculate the midpoint time between two times
}
\examples{
midtime(Sys.time(), Sys.time()-100)
}
