% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_comcat.R
\name{times_to_comcat_url}
\alias{times_to_comcat_url}
\title{Convert a vector of times to comcat_url objects}
\usage{
times_to_comcat_url(times, ...)
}
\arguments{
\item{times}{vector of times; anything that \code{\link{.to_posix}} can handle}

\item{...}{additional parameters sent to \code{\link{make_comcat_url}}}
}
\value{
\code{'comcat_url_list'} object with \code{'comcat_url'} entries
}
\description{
Convert a vector of times to comcat_url objects
}
\details{
Uses \code{\link{.seq_to_seg}} to generate time-segments
}
