% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{time_limit_splitter}
\alias{time_limit_splitter}
\alias{.seq_to_seg}
\alias{midtime}
\alias{is.date_or_time}
\alias{.to_posix}
\title{Create time segments from date pairs}
\usage{
time_limit_splitter(
  now,
  then,
  n,
  paramlist = list(),
  return.list = TRUE,
  verbose = FALSE
)

.seq_to_seg(times, return.list = TRUE, prefix = "seg_")

midtime(early, late)

is.date_or_time(x)

.to_posix(x, verbose = FALSE, to_char = TRUE)
}
\arguments{
\item{now}{a start date; anything coercible to POSIXlt; if missing, \code{\link{Sys.time}} is used}

\item{then}{an end date; anything coercible to POSIXlt; if missing, \code{now - 30} (days) is assumed}

\item{n}{integer; the number of total segments; if missing, this will be larger of either
the number of days between \code{then} and \code{now}, or 3.}

\item{paramlist}{list; parameters, like from a \code{'comcat_url'} object (e.g., from \code{\link{make_comcat_url}}).}

\item{return.list}{logical; should the result be a list? If \code{FALSE} the result is a \code{data.frame}}

\item{verbose}{logical; should messages be given?}

\item{times}{the sequence of times to segment}

\item{prefix}{the prefix to attach to segment identifiers}

\item{early}{the earler time}

\item{late}{the later time}

\item{x}{object}

\item{to_char}{logical; should the final output be a formatted string instead of a
\code{\link{POSIXlt}} object?}
}
\value{
a \code{\link{data.frame}} or \code{\link{list}} depending on the \code{return.list} argument
}
\description{
Create time segments from date pairs
}
\details{
\code{\link{time_limit_splitter}} inspects the contents of \code{params}
for start and end times (in the list with names \code{'starttime'} and \code{'endtime'}); if they exist they will
superceded by non-missing \code{now} and \code{then} arguments.
}
\examples{
L <- time_limit_splitter(verbose=TRUE)
Df <- time_limit_splitter(return.list = FALSE)
Df
all.equal(L, split(Df, Df$Segment))
}
